Some refactoring: packages reorganized.
