+++ src/main/java/chess/game/PieceMover.java
@@ -2,8 +2,7 @@ package chess.game;
 
 import java.util.List;
 
-import chess.Move;
-import chess.pieces.Piece;
+import chess.game.pieces.Piece;
 
 /**
  * Mover is an piece behavior login object, which defines piece behaviour.
